fix: move default toToken logic to selector
